+++ OMG.PSUtilities.Core/CHANGELOG.md
@@ -1,2 +1,9 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.Core
+
+## [0.1.1] - 2025-07-16
+- Find-PSUFilesContainingText : 'Removed a symol in the Write-Warning'
+
+## [0.1.2] - 2025-07-16
+- Find-PSUFilesContainingText : Removed emojies
+- Uninstall-PSUInstalledSoftware : removed symols